### 2020-11-19 #2841 @KristinaRiemer
@dglmoore @marcjwilliams1 @eviatarbach would you be able and willing to review this JOSS submission?
