+++ .github/workflows/Documentation.yml
@@ -1,6 +1,7 @@
 name: Documentation
 
 on:
+  release:
   push:
     branches:
       - master
@@ -8,19 +9,16 @@ on:
   pull_request:
 
 jobs:
-  docs:
-    name: Documentation
+  build:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-      - uses: julia-actions/setup-julia@v1
+      - uses: julia-actions/setup-julia@latest
         with:
-          version: '1.4'
-      - run: |
-          julia --project=docs -e '
-            using Pkg
-            Pkg.develop(PackageSpec(path=pwd()))
-            Pkg.instantiate()'
-      - run: julia --project=docs docs/make.jl
+          version: '1.5'
+      - name: Install dependencies
+        run: julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
+      - name: Build and deploy
         env:
-          GITHUB_TOKEN: ${{ secrets.TOKEN }}
+          GITHUB_TOKEN: ${{ secrets.TOKEN }} # For authentication with GitHub Actions token
+        run: julia --project=docs/ docs/make.jl
