### 2020-11-30 #2872 @whedon
[REVIEW]: SimpleSDMLayers.jl and GBIF.jl: A Framework for Species Distribution Modeling in Julia
**Submitting author:** @gabrieldansereau (<a href="http://orcid.org/0000-0002-2212-3584">Gabriel Dansereau</a>)
**Repository:** <a href="https://github.com/EcoJulia/SimpleSDMLayers.jl" target ="_blank">https://github.com/EcoJulia/SimpleSDMLayers.jl</a>
**Version:** v0.3.4
