### 2020-12-17 EcoJulia/SimpleSDMLayers.jl#41 @marcjwilliams1
JOSS: Downloading data for examples
This is related to openjournals/joss-reviews#2872.

So I've managed to install the package, but running through some of the examples in the docs I get this error. Do you have any idea what might be going? Something on my end or something wrong with the site?

```julia

julia> precipitation = worldclim(12)
[ Info: Downloading wc2.0_bio10m.zip
ERROR: HTTP.ExceptionRequest.StatusError(404, "GET", "/data/worldclim/v2.0/tif/base/wc2.0_10m_bio.zip", HTTP.Messages.Response:
"""
HTTP/1.1 404 Not Found
Date: Thu, 17 Dec 2020 17:42:41 GMT
Server: Apache/2.4.18 (Ubuntu)
Last-Modified: Mon, 12 Oct 2020 23:20:08 GMT
ETag: "df-5b1818a0b093c;5b18143b1d0ec"
Accept-Ranges: bytes
Content-Length: 223
Content-Type: text/html

<!DOCTYPE html>
<html>
<head>
  <meta http-equiv="Refresh" content="5; URL=https://worldclim.org/">
</head>

<body>

<h1>Error Loading Page (404)</h1>

The page or file you are looking for is not available.

</body>
</html>""")
```
