### 2020-12-17 EcoJulia/SimpleSDMLayers.jl#41 @gabrieldansereau
Is it possible you're using an older version of the package? One from the `0.2.x` series perhaps?

It looks like `worldclim()` is attempting to download the WorldClim 2.0 data, but we switched to WorldClim *2.1* starting from SimpleSDMLayers `v0.3.0`. If I remember correctly, there were some changes to the WorldClim website at some point, so we had to change the download URLs accordingly. It's likely that downloads will not work anymore when using previous versions of the package.

So I would recommend upgrading to the latest release, `v0.3.2`, if that's not the case. A few of the features mentioned in our JOSS article were also only implemented starting with `v0.3.0`.
