### 2020-12-18 EcoJulia/SimpleSDMLayers.jl#41 @marcjwilliams1
Yep, that was it. I was still in julia v1.2 which was incompatible with the latest version. Thanks!
