+++ .github/ISSUE_TEMPLATE/to-do.md
@@ -1,0 +1,16 @@
+---
+name: To-do
+about: Short notes on development tasks
+title: ''
+labels: need-triage, to do
+assignees: 'tpoisot', 'gabrieldansereau'
+---
+
+**What to do?**   
+...
+
+**Why?**   
+...
+
+**Any ideas how?**   
+...
