### 2020-12-23 #2872 @marcjwilliams1
Overall I think the package and article are excellent and they more than satisfy all the requirements for JOSS. The documentation and API is clear and easy to use and it is great to see high coverage testing (>80%) of the functionality. A few small comments below:
1. The CI is currently failing on master,I can see previous versions have been passing so have no concerns that these are important issues.
2. Although far from an expert on computational ecology I can see how this package would make some types of modeling easier through integration with other packages as mentioned in the article, are there examples of this in the wider EcoJulia ecosystem? I think it would be nice to mention if so.
