### 2021-01-04 #2872 @KristinaRiemer
@marcjwilliams1 thank you for your review and your thoughtful questions/comments, hopefully @gabrieldansereau will have a chance to respond soon.

@dglmoore, I just noticed that you were not added as an assignee to this issue, I'm going to try to fix that now.
