### 2021-01-05 #2872 @dglmoore
The article and documentation are great. The package has some really nice features that should make many common tasks painless. Performance was good and the code quality looks high.

The only concern at the moment is the failing unit tests, but I suspect that's a simple enough issue to resolve. With this possible exception, all of the JOSS requirements are met or exceeded in my opinion.
