### 2021-01-06 #2872 @gabrieldansereau
@marcjwilliams1 @dglmoore Thank you for your helpful reviews and sorry for the delay.

I fixed the tests in EcoJulia/SimpleSDMLayers.jl#44 and released a new version of the package. As expected by the reviewers, it was a simple fix, a problem with the download URL for one of the supported datasets, which seems to have changed. All other functionalities should work properly.

@marcjwilliams1's comment on integration with other EcoJulia packages is very relevant, however, we do not have such an example of wider integration yet. For now, EcoJulia has a few combinations of packages working together, such as SimpleSDMLayers.jl and GBIF.jl (in review here), or Mangal.jl and EcologicalNetworks.jl (also in review at JOSS). Wider integration between more packages is a long term goal of EcoJulia, but we are not quite there yet. We will for sure include some examples in the package's documentation once we do.
