### 2021-01-25 #2872 @KristinaRiemer
Thanks for the confirmations, @marcjwilliams1 & @dglmoore!

@gabrieldansereau I went through the article proof to check for reference formatting, typos, etc., and found a few things to improve:

- line 53: remove "an"
- lines 129-131: change verbs to clip, coarsen, and perform
- Chamberlain et al., 2020 should have rgbif in title lower case
- add links to software references, e.g., [the suggested citation for rgbif](https://cran.r-project.org/web/packages/rgbif/citation.html) includes link to CRAN page

You should also read through the PDF and make sure you're happy with it, let me know when you've made those updates and are done.
