+++ joss/paper.bib
@@ -9,10 +9,11 @@
   publisher = {{American Association for the Advancement of Science}},
   issn = {2375-2548},
   doi = {10.1126/sciadv.aat4858},
+  url = {https://advances.sciencemag.org/content/5/1/eaat4858},
+  urldate = {2020-03-14},
   abstract = {Demand for models in biodiversity assessments is rising, but which models are adequate for the task? We propose a set of best-practice standards and detailed guidelines enabling scoring of studies based on species distribution models for use in biodiversity assessments. We reviewed and scored 400 modeling studies over the past 20 years using the proposed standards and guidelines. We detected low model adequacy overall, but with a marked tendency of improvement over time in model building and, to a lesser degree, in biological data and model evaluation. We argue that implementation of agreed-upon standards for models in biodiversity assessments would promote transparency and repeatability, eventually leading to higher quality of the models and the inferences used in assessments. We encourage broad community participation toward the expansion and ongoing development of the proposed standards and guidelines. Biodiversity assessments use a variety of data and models. We propose best-practice standards for studies in these assessments. Biodiversity assessments use a variety of data and models. We propose best-practice standards for studies in these assessments.},
   chapter = {Review},
   copyright = {Copyright \textcopyright{} 2019 The Authors, some rights reserved; exclusive licensee American Association for the Advancement of Science. No claim to original U.S. Government Works. Distributed under a Creative Commons Attribution NonCommercial License 4.0 (CC BY-NC).. This is an open-access article distributed under the terms of the Creative Commons Attribution-NonCommercial license, which permits use, distribution, and reproduction in any medium, so long as the resultant use is not for commercial advantage and provided the original work is properly cited.},
-  file = {/home/gdansereau/Zotero/storage/TUVYT3HI/Araujo2019StaDis.pdf;/home/gdansereau/Zotero/storage/KEKQ6PPW/eaat4858.html},
   journal = {Science Advances},
   keywords = {annotated,priority},
   language = {en},
@@ -29,7 +30,8 @@
   pages = {65--98},
   issn = {0036-1445, 1095-7200},
   doi = {10.1137/141000671},
-  file = {/home/gdansereau/Zotero/storage/HEMXXR59/Bezanson2017JulFre.pdf},
+  url = {https://epubs.siam.org/doi/10.1137/141000671},
+  urldate = {2019-04-29},
   journal = {SIAM Review},
   language = {en},
   number = {1}
@@ -44,9 +46,10 @@
   pages = {1--9},
   issn = {1472-4642},
   doi = {10.1111/ddi.12144},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/ddi.12144},
+  urldate = {2019-08-16},
[…]
