### 2021-01-25 #2872 @gabrieldansereau
@KristinaRiemer Thank you for pointing these out.

I made the changes you requested, along with a few minor ones while re-reading. @tpoisot and I are now happy with the PDF.
