### 2021-01-25 #2872 @KristinaRiemer
Great, thank you for doing that @gabrieldansereau. The next steps are to: 1) create a new tagged release on GitHub and 2) archive the code (on Zenodo, figshare, institutional repository, etc.) to generate a DOI. If you can post a link to the DOI here when you're done please.
