### 2021-01-25 #2872 @gabrieldansereau
@KristinaRiemer I've released a new version (v0.3.4). The DOI for the [Zenodo archive](https://zenodo.org/record/4464943) is `10.5281/zenodo.4464943`.
