### 2021-01-25 #2872 @KristinaRiemer
It looks like there's an additional author for the Zenodo archive, should that person be included as an author for the JOSS publication?
