### 2021-01-25 #2872 @gabrieldansereau
@KristinaRiemer No, although he is an important contributor to the EcoJulia organization, he was not involved in this submission to JOSS and his contribution to this specific package was very minor.

I believe he was listed automatically on Zenodo for 2 minor commits in 2019 (one small syntax and one default parameter change).
