### 2021-01-26 #2872 @gabrieldansereau
@KristinaRiemer I believe this is because of the `⋮` and `⋱` symbols on line 161 of the markdown file. I just realized they are not in the PDF either (line 86 of that file).

I'll try removing them as they are not really needed.
