### 2021-01-26 35ea5b1
:wrench: remove problematic symbols for JOSS pdf
