### 2021-01-26 5cc32ca
:wrench: remove problematic symbols for JOSS pdf
