### 2021-01-26 #2872 @gabrieldansereau
Just removed on `master` in [this commit](https://github.com/EcoJulia/SimpleSDMLayers.jl/commit/5cc32ca5241c77a1681ddd07273c75c5dc5fe903).

Do I need to release a new version too?
