### 2021-01-26 #2872 @danielskatz
@KristinaRiemer - feel free to ask @openjournals/joss-eics about things like this, so that you get whichever one of us is currently on-duty (me, this week).

No, a new release is not needed for small changes in the paper.  Just run the accept again...
