### 2021-01-26 #2872 @gabrieldansereau
Could the error happen because I had `bibliography: [paper.bib]` instead of `bibliography: paper.bib` in the markdown header? I made the change [here](https://github.com/EcoJulia/SimpleSDMLayers.jl/commit/789e8134de5c2a90ffba8d0a1343d7045fba48ee), maybe we should try again? (It still works with the JOSS preview service)
