### 2021-01-26 #2872 @arfon
`bibliography: paper.bib` should fix the paper compilation issue here.
