### 2021-01-27 #2872 @danielskatz
@gabrieldansereau - the zenodo archive has a different title and authors than the paper. Usually, these should be the same. Can you change the zenodo metadata or explain why these are different?
