### 2021-01-27 #2872 @danielskatz
You can (and I think should) change the zenodo metadata to match the paper - you don't have to keep the automatically compiled metadata.
