### 2021-01-27 #2872 @tpoisot
@danielskatz - I think the contributions to the code and paper can be kept different, and I would not be comfortable changing the metadata on zenodo.
