### 2021-01-27 #2872 @danielskatz
I've also suggesting changing some cases in the paper text: https://github.com/EcoJulia/SimpleSDMLayers.jl/pull/46
