### 2021-01-27 #2872 @tpoisot
the casing changes have been merged
