### 2021-01-27 #2872 @danielskatz
> I would not be comfortable changing the metadata on zenodo.

Why not? This is a normal thing that is done by JOSS authors. We really would like you to change the title, at least, as the automatically generated title is not very descriptive, though maybe leaving the authors as it is makes sense in this case.
