### 2021-01-27 #2872 @danielskatz
The paper title is: SimpleSDMLayers.jl and GBIF.jl: A Framework for Species Distribution Modeling in Julia
We normally expect that for a JOSS paper there is an archive that contains a snapshot of the software that was reviewed, and the title of that archive entry matches the title of paper.

In the paper draft, I see a link to a zenodo archive title with the title: EcoJulia/SimpleSDMLayers.jl: v0.3.4
Normally, this would be an archive for all the software that was reviewed.
Is there also a second zenodo archive for GBIF?

If so, I think perhaps we want a single archive that contains both, which in this case will have to created manually, since JOSS does not have a way to link to two archives.

And maybe @arfon will have some additional thoughts on this
