### 2021-01-27 #2872 @tpoisot
The dual archive would not work with Julia packaging system, which requires the packages be kept in separate repos. There is, of course, a second zenodo archive for GBIF.jl - https://zenodo.org/record/4293898#.YBGAZtZyYew - making a single archive for the two of them would lead to duplication of content and ensure that the joint archive is not kept up to date, and it's not me being obtuse, I just don't see the benefits for users.

I'd also like to add that GBIF.jl is not *necessary* for for SimpleSDMLayers.jl to function (and that wen installed, only the version of it compatible with SimpleSDMLayers.jl will be available.
