### 2021-01-27 #2872 @danielskatz
In a JOSS paper, there is a single link that should point to an archive of all the reviewed software.
(This archive is a permanent record of the reviewed software, not a live version.)

In this case, the link does not - it only points to part of the reviewed software. To me, this is a problem.
But it's easy to fix, by creating an archive (in zenodo, figshare, etc.) that contains both packages - again, this is for the scholarly records associated with this paper, not for people who want to use the software, who would go to the github repos.  And yes, this has to be done manually in this case, because it's not something the zenodo-github linkage supports.
