### 2021-01-27 #2872 @danielskatz
> ensure that the joint archive is not kept up to date, and it's not me being obtuse, I just don't see the benefits for users.

This doesn't need to be kept up to date, and it's not for users.  It's for the scholarly record associated with this paper, and is a one-time need for the paper.
