### 2021-01-27 #2872 @tpoisot
Is it better than linking both DOIs? I can do a one-time archive with both. if this is a _requirement_ from the editorial board, althought I don't see the point. Just tell me which title you think it should have and I will get it done.
