### 2021-01-27 #2872 @danielskatz
Linking both DOIs would be better in some ways, but the JOSS paper template doesn't support it.  There is a requirement to link to the reviewed software, and I think because of the JOSS template, we need a single archive that contains it.
