### 2021-01-27 #2872 @KristinaRiemer
Looks like whedon was a little slow, does [this](https://github.com/openjournals/joss-reviews/issues/2872#issuecomment-768377632) need to be dealt with?
