### 2021-01-27 #2872 @arfon
> Linking both DOIs would be better in some ways, but the JOSS paper template doesn't support it. There is a requirement to link to the reviewed software, and I think because of the JOSS template, we need a single archive that contains it.

:wave: folks - just popping in here to say that currently this is the only way for JOSS to handle submissions like this. I would like us to be able to link to multiple software archives (and repositories) but the current model for JOSS is one review per software repository.
